fix nova path and move single file dirs down
